server/requirements.txt
Django==1.5.4
Fabric==1.6.1
South==0.8.1
boto==2.9.5
distribute==0.6.36
dj-database-url==0.2.1
dj-static==0.0.5
django-appconf==0.6
django-cache-machine==0.8
django-compressor==1.3
django-discover-runner==1.0
django-extensions==1.1.1
django-grappelli==2.4.5
django-redis-cache==0.9.7
django-storages==1.1.8
django-tastypie==0.9.15
django-proxy==1.0.2
gunicorn==0.17.4
hiredis==0.1.1
mimeparse==0.1.3
paramiko==1.10.1
path.py==4.1
psycopg2==2.5
pycrypto==2.6
python-dateutil==2.1
redis==2.7.5
requests==1.2.3
simplejson==3.3.0
six==1.3.0
static==0.4
wsgiref==0.1.2
